// src/core/enums/production-establishment.enum.ts
export enum ProductionStablishmentEnum {
  Daily = 1,
  Weekly = 2,
  TwoWeek = 3,
  Monthly = 4,
  FortyFiveDays = 5,
  SeventyFiveDays = 6,
  ThreeMonth = 7,
  Seasonal = 8,
  FourMonth = 9,
  FiveMonth = 10,
  SixMonth = 11,
  EightMonth = 12,
  NineMonth = 13,
  TenMonth = 14,
  ElevenMonth = 15,
  OneYear = 16,
}


export enum ProductionStablishmentPersianEnum {
  Daily = "روزانه",
  Weekly = "هفتگی",
  TwoWeek = "دو هفته ای",
  Monthly = "ماهانه",
  FortyFiveDays = "45 روزه",
  SeventyFiveDays = "هفتاد و پنج روزه",
  ThreeMonth = "سه ماهه",
  Seasonal = "فصلی",
  FourMonth = "چهار ماهه",
  FiveMonth = "پنج ماهه",
  SixMonth = "شش ماهه",
  EightMonth = "هشت ماهه",
  NineMonth = "نه ماهه",
  TenMonth = "ده ماهه",
  ElevenMonth = "یازده ماهه",
  OneYear = "یک ساله",
}